Update patch set 1
Patch Set 1:

My patch can be reverted for beta 3 if it cause troubles. It is mostly used on RISC-V port where graphics card is not initialized by BIOS/firmware and app_server initializes graphics card for the first time. It is not very relevant for x86 platforms, video card is usually initialized by BIOS and video BIOS there.

Patch-set: 1