Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 2:
> > > 
> > > > It is already fixed upstream.  Should we just copy their fix for this?  
> > > 
> > > Yes.
> > 
> > Copied fix from upstream.
> Ugh, the online editor formatted text (indentions, etc.) doesnt match what shows on the diff.

It's not recommended to use it except for very simple changes.

In particular I don't think it's possible to set the correct tab size in the online editor.

It's better to do changes locally, amend the commit and push it again (with the same change-id in the commit message).

Patch-set: 3
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>