Update patch set 1
Patch Set 1:

How so? It checks for overflows and then IS_USER_ADDRESS on both ends of a range, just like this does. This needs an extra check because it needs the address to be a user one, but otherwise that function should be used here.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>