Update patch set 1
Patch Set 1:

The original intent of this was to make this class more generic for use in other places.

I think we have settled on using SettingsMessage for settings (now available in libshared), and so this patch is not useful anymore. Should we abandon it?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>