Update patch set 2
Patch Set 2:

The other alternative is to not return an error, as the package kit detects if it had to create the administrative folder, and runs the first boot thing anyway. But creating the directory doesn't hurt, so I think this is a better option.

See: https://github.com/haiku/haiku/blob/master/src/apps/installer/WorkerThread.cpp#L319

Patch-set: 2