Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> I meant "Spacing" as it functions in FontDemo.
> 
> But I agree it works well enough for the common use case.

The spacing in FontDemo is implemented by drawing characters one by one with a separate call to DrawString for each character, giving a BPoint for each character. So it seems normal that this does not result in a continuous line.

I am trying to modify it to use a single DrawString with positions for each character. In that case, there is a continuous underline even when changing the spacing, but it isn't affected by the font rotation. And there are some other problems with using DrawString this way in FontDemo anyway. I will see what I can do about them.

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>