Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> That fixes the crash for me. Would be nice to have in the beta...
> 
> It's not perfect though:
> 1) The menu has a nervous twitch, every Pulse() it gets a visible redraw.
> 2) Every redraw it loses the selection of the item under the mouse pointer. Gotta keep that pointer moving...

Visible redraw was present before and is indeed annoying. I tried turning updates off and on but that  didn't fix it.
Losing selection under item was present before. I had fixed this in earlier version but I had to extend BMenu to do it.

I'd like to do a second version of patch that does more of the work like sorting and selecting in BMenuPrivate extensions. This version of patch is intentionally isolated so as to hopefully get a fix into beta 3 that fixes the crash before addressing these other issues.

Patch-set: 1