Update patch set 5
Patch Set 5: Code-Review-2

The general opinion on the forums seems to be: do not use third-party services without user consent, and especially not for location data. I tend to agree, and so until we have reached a decision or consensus about that, this shall not be merged.

Patch-set: 5
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2