Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I think that this path is conceptually good, but BGeolocation implementation should ask user confirmation if it use access to external servers and send user WiFi information. If BGeolocation is based on GPS, it should be fine to use, no data is sent.

In this specific case, if asking the user anything is needed, I think I would rather do nothing. The idea is to reduce the effort to pick the correct language from 1 click to less than 1 click. If we need an extra confirmation, it makes it 1 click or more: if you don't allow geolocalization it's two click ("no" + the correct language). If you allow it, it's at least one click, maybe two if it doesn't work (network offline, or it just pick the wrong language).

Also, the confirmation would be asked in english, which defeats the whole purpose of this.

So, I think what we need to do this is add some flags to BGeolocation to do a geolocation without leaking any data. And since GPS is not that common in desktop computers, is it still worth the effort?

My backup plan is looking at the language field in HID report of USB keyboards (not for geolocation, but for guessing a language to use). But I think only Apple keyboards actually implement it.

Patch-set: 5