Update patch set 1
Patch Set 1:

I think a nice follow-up patch would be to only recreate the menu when there were any changes. Since this will happen very rarely, the repainting issue would be gone completely, and the select issue would be way less annoying -- and even not a bad idea, because it would be kinda surprising if the item under the mouse changes a fraction of a second before you click on it.

Patch-set: 1
CC: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>