Update patch set 1
Patch Set 1:

Oh, so you guys want a full solution with cherry on top... 😊:)

That'll entail some sort of NodeMonitoring, I suspect, which I'd have to read up on and lead to a fair amount of experimenting. I also don't have a CD drive. And foremost, not much time lately.

I'll have to abandon this patch, I think. Giving it a bit more time for potential feedback.

Patch-set: 1