Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-2
> 
> Is this a change for change sake? Is the function used elsewhere?

It's not about the function per se. The function takes a reference to a class/struct.
The function is declared just outside of an #ifdef __cplusplus block, when it should be inside that block. See patchset 1, for a less invasive fix. (Just moves the declaration two lines).

I don't recall now, why I was including platform.h from a plain old c file. (probably wanted dprintf()), but why not fix it?

Patch-set: 3