BMenu: Add SortItems(), SwapItems() and MoveItem()
... methods which call the respective methods in BList.

These methods are public but not virtual so they don't change the
size of the class. The methods are non-virtual in BList as well.

These convinience methods allow you to sort a menu of menu items
via a compare function, swap two menu items, or move a menu item
to a new index.

Previously there was no easy way to rearrange menu items in a menu.