Update patch set 2
Patch Set 2:

(2 comments)

Although the previous pointer comparison code worked, it was probably not reliable. I overhauled how I verify the equality of wireless_network structs by creating an == operator method. BNetworkAddress is not considered for equality but the rest of the struct fields are. There doesn't seem to be a reliable way to compare structs for equality other than to compare their values, and my attempt to compare pointers was futile.

Patch-set: 2