Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> How updating is supposed to be handled?

I assume by updating you mean overriding. Simple answer is that it's not. That's the question, is that ok?

If I have to make these methods virtual then it will seriously cut into the remaining virtual function FBC reserve to preserve binary compatibility (which is 4 for BMenu.)

Patch-set: 1