Update patch set 1
Patch Set 1:

The simplest solution to adding these methods privately is to put them in BMenuPrivate class. I'd have to update the #12024 patch to call the methods in BMenuPrivate too. If we need these methods to be virtual that's probably the way to go because there are only a few FBC class methods left in BMenu.

Patch-set: 1