Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I hope that the public scope of these methods does not offend.
> > 
> > I hope that the non-virtual status of these methods does not ruin polymorphism.
> > If these methods should be virtual, they should be virtual in BList too.
> 
> I don't see why one would want to override these methods, and I see no reason to make them private either.

Ok good that's what I was afraid of.

> The question is more, do they work reliably? Can we call them while the menu is open, for example? What happens then?
> 
> What happens if I add a menubar to a window, then call SortItems on my menubar? Does it update on screen correctly or do I need to manually Invalidate it? What if I am a bit crazy and do this while a menu with multiple submenus is open? Should things move immediately? Should the change wait until the menu is closed?

These are good questions that I don't have good answers for.

I only tested SortItems() in the Network preflet. It flashes still on Pulse(), but, because the menu items remain in the same location in the menu, the flashing is not nearly as bad as it was before.

Patch-set: 1