Update patch set 1
Patch Set 1:

B_NO_COMMAND_KEY is better. I made my own version of this patch as well, but never quite got it to work. I want to change WebPositive back shortcut from Command+left to Control+left. Why? So that it will be the same as Windows except with Ctrl and Alt reversed. The shortcut needs to appear in the menu to be done right.

Patch-set: 1
CC: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>