Update patch set 1
Patch Set 1: Code-Review-1

This looks like a mess. You are mixing generic ECAM stuff that should be machine-independent with the platform-specific FU740 PCIe controller (i.e. the former is incorrectly duplicated into this architecture-specific directory, and the latter is incorrectly in a file that's not platform-specific enough, they're both in the wrong place and should never be in the same file). There is no such thing as a riscv64 PCI controller.

Patch-set: 1
Reviewer: Gerrit User 1000281 <1000281@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1