Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> BBL is also considered obsolete. I am not sure can latest version of OpenSBI be used with TinyEMU. TinyEMU is currently fastest RISC-V emulator for me, QEMU is slower. TinyEMU is also have small code size and it is easy to understand.

BBL does still work. There should be no reason why OpenSBI wouldn't work but BBL would, it provides a strict superset of the functionality at this point.

TinyEMU is non-conforming in ways that OSes care about; I know of bugs in its PLIC implementation and in its VirtIO block device implementation, both of which break FreeBSD. But feel free to use it if you want, just use BBL or OpenSBI like you're meant to if you want to support it in Haiku, and bear in mind that it's a bit buggy in places (unless that's been fixed since I looked at it a bit over a year ago), because nobody in the RISC-V community is going to endorse running an OS bootloader in M-mode.

Patch-set: 1
Reviewer: Gerrit User 1000281 <1000281@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>