Appearance: Store decorator name to disk untranslated
Save the path of the Default decorator as "Default" to disk because
that's the string app server is looking for to identify it.

We now save all decorators names to disk untranslated. We no longer
translate strings in DecorInfo class, only in Apperance Prefs.

Use BPrivate::gSystemCatalog in Appearance to translate "Default"
decorator name so that we don't have to retranslate the string
e.g. Predeterminado.

Fixes #16412

Change-Id: I3109681d11931e7f55b7afaf3a65af2fb9ed5e10