Appearance: Store decorator filenames to disk untranslated
Save the ShortcutName() which is the filename of the translator
instead of the potentially translated decorator Name(). If the
decorator has no ref but fInitStatus is B_OK return "Default"
for default decorator in ShortcutName(). The Default decorator
unlike other decorators has no file on disk associated with it.

We don't need to look for ShortcutName() when iterating through
decorators anymore because we look for "Default" then file names
before we try to find by name.

Fixes #16412

Change-Id: I3109681d11931e7f55b7afaf3a65af2fb9ed5e10