Update patch set 4
Patch Set 4:

I'm not actually sure how rarely FPU is used in the kernel, to be honest. It is a comparatively small amount of code, but some of it might be in the critical paths in some places. I think on x86 we don't do lazy FPU switching (but I think there were also some Spectre related vulnerabilities to doing it anyway?)

Patch-set: 4
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>