Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> We definitely want to share code between kernel and bootloader where possible. (Reducing bootloader surface is also good; I think x86 uses some tricks here to set up the absolute minimum of pagetables before the kernel takes over, and probably RISCV could do similar.)
> 
> Also, thanks Jessica for reviewing this stuff; advice from experts is always invaluable. (How did you come by Haiku, anyway, if you don't mind me asking? :)

I was helping kallisti5 on IRC understand how page tables work, specifically in the context of RISC-V in order to get your EFI bootloader working on it, and reviewed various versions of that patch on here to point out issues. That patch was superseded by one of X512's in this series the other day so I skimmed the titles of the patches and figured I'd look at the ones that sounded more technically interesting and give feedback as someone who has a lot of experience with RISC-V and the HiFive Unmatched/FU740.

Patch-set: 1