Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > as someone who has a lot of experience with RISC-V and the HiFive Unmatched/FU740.
> 
> Can you tell does HiFive Unmatched have Platform DMA controller? I plan to use it for memcpy and video memory copy because CPU memcpy is terribly slow (~120 MB/s).

Yes, the Unmatched uses the FU740 and the FU740 includes a PDMA controller at 0x300000 (called just "DMA" in the memory map) with 4 channels. 120 MiB/s sounds slower than I remember, are you sure you were doing full words at a time? But also, the latency for setting up DMA transfers and waiting for completion will make it not worthwhile for almost all memcpy uses, and it'll only be usable in the kernel. Though I don't know how much faster the bandwidth will be. Video memory is about the only thing it might make sense for. Note also there is not yet a device tree binding specified for it, let alone a driver for any of FreeBSD, Linux and U-Boot.

Patch-set: 1