Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> For my RISC-V work bare metal boot loader is essential part. It is used to quickly boot system on virtual machines. This helps testing a lot. It will be harder to me if haiku_loader.riscv will bd removed from tree.
> 
> With haiku_loader.riscv boot loader finishes loading kernel with less than 0.5 second. With haiku_loader.efi and u-boot boot loader takes several seconds because u-boot is bloated.

I already told you, if you use OpenSBI directly instead then you get almost exactly the same speed, it only takes a fraction of a second itself. That is what you're meant to use if you want to avoid EFI (and is what was always used before EFI was implemented for RISC-V anywhere). It is not a binary "M-mode or EFI" choice.

Also, for what it's worth, the delay with U-Boot you see on QEMU is *not* because of bloat. It is because it defaults to waiting a few seconds before booting so that you can drop to a shell. If you configure U-Boot's bootdelay to be 0 then you can skip that.

So, no, you are completely wrong here. Please stop pushing this argument, it's tiring. If you want to argue for supporting M-mode because you want to then fine, but it serves no practical purpose other than that (not even "not having to build firmware", because a pre-built OpenSBI binary ships with QEMU).

Patch-set: 1
Reviewer: Gerrit User 1000281 <1000281@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>