Update patch set 1
Patch Set 1:

Hello,

If you want to continue arguing over this, I would appreciate if you could move the discussion elsewhere where I don't get email notifications for every message.

I think everyone agrees that this machine mode loader will not be distributed with Haiku and used on real hardware. But it seems to be an interesting experiment and possibly useful tool for some workflows (wether the workflow in question is "right" or "wrong" is not our problem).

Personally I'm fine with merging this code in Haiku, knowing that we won't in fact use it for the final build images which will use existing firmware. I don't think it gets in the way of anything or anyone else?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>