Update patch set 1
Patch Set 1:

> If you want to continue arguing over this, I would appreciate if you could move the discussion elsewhere where I don't get email notifications for every message.

We should probably preserve this argument for posterity. Where else should it be held? (Can you just unsubscribe from this one change?)

> I think everyone agrees that this machine mode loader will not be distributed with Haiku and used on real hardware. But it seems to be an interesting experiment and possibly useful tool for some workflows (wether the workflow in question is "right" or "wrong" is not our problem).

Jessica Clarke's point appears to be that there is a mechanism (OpenSBI) by which nearly all the technical advantages X512 wants, namely rapid boots without the "bloat" of EFI et al., can be had, with none of the downsides.

Further, there is a reason EFI takes longer to boot: it is my understanding that EFI takes care of assigning a number of things that you have to do yourself in bare metal or SBI mode. That is another reason we may not want this code in tree to have to deal with it.

> Personally I'm fine with merging this code in Haiku, knowing that we won't in fact use it for the final build images which will use existing firmware. I don't think it gets in the way of anything or anyone else?

It depends. If whatever EFI takes care of extends into what the kernel has to deal with, then "supporting" this boot mode is not just code stuffed in a little-used bootloader, but may affect kernel behavior or at least code behind #ifdef's. That would be a major reason not to use it.

Patch-set: 1