Update patch set 1
Patch Set 1:

> This discussion cause me to start thinking to abandon attempts to work with Haiku source tree and Gerrit for RISC-V port and use my own fork instead.

I and the other project developers and contributors see value in consensus, not mere tolerance of varying opinions, and discussions like these (despite the fact that Jessica is not a "Haiku developer" as such) are precisely how such consensus are built; so if your thoughts are coming from the long-windedness of the discussion itself and not the content thereof, I would encourage you to try and see why a long argument is not necessarily a bad thing.

I would further argue that, if your major disagreement is instead with the technical content of the discussion, that even if some consensus is reached among other developers and contributors that you ultimately disagree with, that going along with it anyway despite disagreement can be good, too. (I realize that many people, of which you may or may not be one, who have to do this in their day jobs find it unappealing to do the same in open-source work, but, well, there is a reason it is a good in general.)

Patch-set: 1