Update patch set 2
Patch Set 2:

Default decorator name and description are still localized. If you are using a decorator with a localized name other than the Default one you may have to reselect it in Appearance. This is however, unlikely. Decorators are now always identified by their untranslated English name but the name is localized before it is displayed in the pop up menu.

In short everything should work as before except it will also select the localized Default decorator now (fixing #16412.)

Patch-set: 2