Update patch set 1
Patch Set 1:

> > Personally I'm fine with merging this code in Haiku, knowing that we won't in fact use it for the final build images which will use existing firmware. I don't think it gets in the way of anything or anyone else?
> 
> It depends. If whatever EFI takes care of extends into what the kernel has to deal with, then "supporting" this boot mode is not just code stuffed in a little-used bootloader, but may affect kernel behavior or at least code behind #ifdef's. That would be a major reason not to use it.

It does not seem to be the case in this patch. It's just some definitions of registers from machine mode in a header file, that has no impact on anything else. It's just a few definitions in a header file.

So, in this specific patch, I see nothing that justifies not merging the changes.

> I and the other project developers and contributors see value in consensus, not mere tolerance of varying opinions, and discussions like these (despite the fact that Jessica is not a "Haiku developer" as such) are precisely how such consensus are built

I think this discussion is not getting anywhere, I see two persons repeating their own arguments over and over again and making no progress to build a concensus. I think that is frustrating for everyone and getting nowhere. Which is why I suggested some middle ground, which I think everyone can agree on? Keep the code for reference and possibly for the specific use case X512 wants it to, yet remember that this will not be used in release images, where we will probably go with relying on the available firmware on real hardware.

I can add an extra thing to this if you're worried it will be polluting the kernel with more specific code: if this gets in the way of simplifying the boot process and adds a lot of code, we can always remove it later when we have a more stable RISC-V support, and everyone is running on real hardware and TinyEMU is not needed anymore.

Patch-set: 1