Update patch set 20
Patch Set 20: -Code-Review

> Patch Set 20: Code-Review+1
> 
> I first thought something's amiss when all 8 cores pegged when I opened a folder with image files, and 'un-pegged' when closing the folder window. I didn't see any thumbnails though...
> But now they're there. I suppose Tracker just crawls all disks the first time to generate them?

I was seeing the same thing where it was taking an obscenely long time to generate the thumbnail at first and spins its wheels trying. But then after that initial delay it generates the thumbnails quickly from then on.

That’s initial delay is not suppose to happen and there may be something blocking it from writing the attribute so it’s generating the thumb over and over again. It seems to only happen on 64-bit Haiku. So we may not be quite there yet but we are getting close.

> 
> One thing: I have some AVI files of videos taken with a camera. Those video files don't have a thumbnail icon, but the normal video file icon. There's a similarly named .THM jpg image alongside them though.
> Is it possible to change the video file's icon to that thumbnail?
> Plus, overlay something to mark it as video, maybe a film strip perforation along the sides, or an icon in the corner?
> 
> Maybe after this patch got merged though...

Assuming that we have a translator for AVI files it should be possible to generate a thumbnail from the video data directly but I’m only targeting images currently. It will probably take some additional work to get video thumbnails to work right but in theory yes we can do that.

Patch-set: 20
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0