Update patch set 20
Patch Set 20:

> Patch Set 20: Code-Review-1
> 
> Pegging all 8 cores does not sound like a good behavior... Can we limit thumbnail processing to maybe 1 or 2 threads (or maybe, 25% of cores in the system), and at a low-priority, too?

The code already spawns a single thread at B_LOW_PRIORITY. At least in theory.

Note that even at low priority, it will use 100% of a CPU core if available, and if the IO device can handle it. There is no reason to make it slower than that. If the CPU is not needed for anything else, it may as well generate thumbnails?

Patch-set: 20