Update patch set 22
Patch Set 22:

> Patch Set 22:
> 
> I only have a 64bit dev environment around. At least on 64bit it seems to work nice and snappy, no more cpu pegging.
> 
> BTW, don't you have to run with the generated libbe.so as well?
> I just put libbe.so and libtracker.so in a "lib" subfolder alongside the generated Tracker binary.

You do have to copy libbe.so for changes to kits, but in this case only Tracker (and libtracker.so) has been changed, and there are no changes to libbe.so. You can put libbe.so in your lib dir but it is not necessary here.

Patch-set: 22
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>