Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Thanks for trying my patch John! To get the new save window up, either close StyledEdit using the "Close All" button in the Deskbar, or attempt to shutdown the system with unsaved StyledEdit files still open. Additionally, I could not seem to reproduce your crash after having multiple unsaved windows open, and then trying to open up a file from StyledEdit's menu. Are there anymore specific details you could give about your test case?

I now see the Save all window, very nice.

Here's the best instructions I have to reproduce the crash. Open a window, make changes, open another window, go back to first window and close it, Don't save. Then go back to the open document and open another window. Keep doing that and you should hit the crash sooner or later. It may not happen every time but it should be fairly easy to get to the crash. This method may not reliably reproduce the crash every time but it's not super hard to do. I haven't quite gotten the exact steps down.

The crash seems to happen with this patch but does not happen with the system StyledEdit (so it's not a pre-existing bug.)

Patch-set: 2