Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Yes, but the point is that FreeBSD only enables/disables SMAP inside the copyin/copyout functions, it appears, not across other parts of the kernel as well.
> 
> IMO adopting that sounds like a good idea. I think SPARC has even more restrictions on accessing user memory than RV64 does, so it may be worthwhile anyway.

Sparc is different, there is no in/out mechanism and the kernel and address spaces are completely separate. To access userspace from kernel, specific instructions need to be used (see "alternate address spaces" or ASI).

The difficulty there will be that there is no way to tell if a pointer is user or kernel side just from the address (they are two completely separate address spaces, so the same address can be valid in both), so IS_USER_ADDRESS/IS_KERNEL_ADDRESS can't be implemented. I think this can be a problem for user_memcpy (which detects which side is userspace and which side is kernelspace) and that's why other OS implement separate, explicit "copyin" and "copyout" paths?

Patch-set: 5
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>