Update patch set 7
Patch Set 7: Code-Review-1

It seems the point of this update was to split some parts of the changes (now in the next commit in this branch) which can be merged separately. This part is still not to be merged.

Patch-set: 7
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1