Update patch set 7
Patch Set 7: Code-Review+1

The developer consensus may still be to have this patch not merged but I'd still like it to be merged anyway so that I can stop using the Scroll Bars preference app from BeOS to change this setting. Two other people bitigchi and CodeForEvolution have expressed support for this being merged in the comments for #9137 and several more people including Disreali, michaelvoliveira, Pete, khallebal and stippi have expressed support for this to be merged in #6206. I feel that there is a divide between the Haiku developers and users regarding this change.  There is a screenshot linked above showing what this change would look like using HaikuDecorator and BeDecorator if you'd like a better idea of what this change entails.

That said I'm not asking for this to be merged at this time, the reason that I created this new patch revision was to split out the 2x thumb dot change in HaikuDecorator. Nothing else has changed since the last revision. I'll provide a screenshot on the other patch set showing the 2px dot change when I get a change.

Patch-set: 7
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1