Update patch set 7
Patch Set 7:

> Two other people bitigchi and CodeForEvolution have expressed support for this being merged in the comments for #9137 and several more people including Disreali, michaelvoliveira, Pete, khallebal and stippi have expressed support for this to be merged in #6206. I feel that there is a divide between the Haiku developers and users regarding this change.

What I remember is that there is general agreement that a setting for this is not needed, but we are open to changing the default look and having scrollbar knobs on by default. But it's been a while, and I could be wrong about that.

Patch-set: 7
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>