Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> What I remember is that there is general agreement that a setting for this is not needed, but we are open to changing the default look and having scrollbar knobs on by default. But it's been a while, and I could be wrong about that.

No I don't think that's quite right. It's that some people are against the idea of having scroll bar knobs at all and some are for it, and the Appearance setting is the sticking point for that divide.

A long time ago a version of this change was briefly committed before being reverted. Back then dots were the the default setting so suddenly people were seeing dots on their scroll bar thumbs and were reacting unfavorably to the change.

The reason that the dots were showing up on peoples scroll bar thumbs back then is that the default value for the scroll bar knob setting was 1 indicating dots style as that was the default scroll bar  knob setting in BeOS and we simply inherited that setting in Haiku.

Since that time I have altered the default scroll bar knob setting in Haiku to be 0 so that you are not opted into scroll bar dots by default, you have to go to Appearance preferences and change the setting to get dots or lines if you want them.

Patch-set: 7
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>