Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What's the correct way to refer to this application in commits?
> 
> * Preferences/Time:
> * Time:
> * Time Preferences:
> 
> Although I'm mostly asking because I just wrote in a fresh About Window while I was looking at whether I could do something more useful, this also concerns this change.

I would say Time Preferences would be the preferred language to use. Although that's only necessary if there is ambiguity and in this case there is not so simply "Time" would also suffice.

I guess you could confuse "Time" with the Clock demo app so "Time Preferences" clears it up, but this is probably not a big concern.

Patch-set: 2