Update patch set 3
Patch Set 3:

I did what you suggested and localized the Short description in DecorInfo again which fixes the immediate problem of that string not being included in the system catalog.

However I feel that this solution is not 100% correct as the Short Description is then inserted into Long description in Appearance and then the Long description is localized. I could imagine a scenario where the localized Short description gets inserted into the non-localized Long description and then when Long description is localized it doesn't find a match in the catalog because of the already localized Short description inside Long description that it's not expecting.

Could we include Short Description in Appearance somewhere so that we can localize the string there instead of in DecorInfo? We could grab the localized Short description from the System catalog in the case of the Default decorator.

This patch also attempts to B_TRANSLATE the rest of the decorator and control look names and put them in the Appearance catalog, not system catalog. Is it wise to attempt to localize the non-Default decorator and control look names considering that they are optional and may or may not appear on any given system. Examples of non-Default decorator and control look possibilities currently include "Be", "Mac", "Flat", and "Windows". I'm not sure if those strings will ultimately make it into the catalog to be localized and I'm not sure if we would want to localize them anyway... maybe just "Flat".

Does it make sense for us to attempt to add non-Default decorator Name, Short description, and Long descriptions to the system catalog instead of using the Appearance prefs catalog?

Patch-set: 3