Update patch set 3
Patch Set 3:

> Does it make sense for us to attempt to add non-Default decorator Name, Short description, and Long descriptions to the system catalog instead of using the Appearance prefs catalog?

Neither.

The add-ons should each have their own catalog with the strings stored there.

The reason for doing it this way is that there are (or could be) decorator add-ons developed by people outside of Haiku. These should have the option to localize things as well, and if the strings are in appearance preferences or the system catalog, they just can't.

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>