Update patch set 3
Patch Set 3:

> How do I add name, short and long descriptions to a catalog for each decorator add-on? Do I need to edit the BDecorator and BControlLook classes? BDecorator [0] and BControlLook [1] currently have no concept of short and long description. 

The way it works currently is that the DecorInfo class is created from resources of the add-on. This would need to be changed so that the add-on has a "DecorInfo* get_decor_info()" function and creates the DecorInfo instance with translated strings.

I thought this was already the case, but after taking a closer look at the code, I see it isn't.

The dowside of this is that it forces us to load the add-ons so we can call this function, just so we can get their name and other info to show in appearance. With the current design, we don't need to do a load_add_on for that, just scan the BResources of the add-on.

> How do I import those catalog files into Appearance prefs so that I can read the localized strings? 

I would let the add-ons fill the DecorInfo class with translated strings (using the usual B_TRANSLATE calls). Appearance would not need to do anything, as it would receive already translated strings, ready to display.

If there is something that makes loading the add-ons and calling a function a bad idea, we would need to have Appearance load the add-on's catalog without loading the add-on itself, and fetch some well-known strings from there. This is possible but I don't think we have any place where we actually did something like that.

It would go this way:

- Find the mime type of the add-on
- Load the catalog matching this MIME type (there is an example of this in Cortex, where a catalog is shared by multiple add-ons and they load it this way)
- Access the strings in the catalog. Normally the index for a catalog is the english string to translate, but here that would not be the case. We have some key ("short description", "long description", ...) and we want the corresponding string. There is a BCatalog::GetString version which takes an uint32 key as an argument, but it was not used in Haiku so far and I think the current catalog format doesn't know how to store something like that, or even if it does, it probably wouldn't work well with Pootle and other tools.

So it turns out to be not as simple as I thought at first...

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>