Update patch set 4
Patch Set 4:

Patch set 4 doesn't attempt to translate third party decorators, instead it stores to disk and later finds the decorator a different way using ShortcutName() instead of Name() which sidesteps the issue for now fixing #16142. ShortcutName() returns the decorator's file name except in the case of the default decorator "Default" is used instead.

Translating the name and short description strings of the default decorator have been restored in DecorInfo and are used by Appearance prefs to show the translated decorator name in the pop up menu.

Patch-set: 4