Update patch set 1
Patch Set 1:

> I think that this will undo Ticket 17078

There is an extra bugfix (the removal of a re-declaration of "status_t status" and we think it could fix that problem. Which is why I asked mmu_man to confirm this.

Also, please write ticket numbers like this: #17078 so that Gerrit links us to the right place 😎

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>